perf(plugins): avoid logging imports during provider discovery
