+++ src/plugin-sdk/diagnostic-flags.ts
@@ -1,0 +1,2 @@
+// Flag checks avoid initializing diagnostic events, tracing, and redaction.
+export { isDiagnosticFlagEnabled } from "../infra/diagnostic-flags.js";
